rename chatTrigger dir to twitchCommands
