--- src/main/talium/system/chatTrigger/cooldown/CooldownType.java
@@ -1,9 +1,0 @@
-package talium.system.chatTrigger.cooldown;
-
-/**
- * A Unit of a Cooldown, used in {@link talium.system.chatTrigger.cooldown.ChatCooldown} to build a value with a unit
- */
-public enum CooldownType {
-    SECONDS,
-    MESSAGES
-}
