+++ src/main/talium/system/twitchCommands/cooldown/CooldownType.java
@@ -1,0 +1,9 @@
+package talium.system.twitchCommands.cooldown;
+
+/**
+ * A Unit of a Cooldown, used in {@link talium.system.twitchCommands.cooldown.ChatCooldown} to build a value with a unit
+ */
+public enum CooldownType {
+    SECONDS,
+    MESSAGES
+}
