+++ src/main/talium/system/twitchCommands/persistence/MessagePatternRepo.java
@@ -1,4 +1,4 @@
-package talium.system.chatTrigger.persistence;
+package talium.system.twitchCommands.persistence;
 
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
