+++ src/main/talium/system/twitchCommands/persistence/MessagePatternService.java
@@ -1,4 +1,4 @@
-package talium.system.chatTrigger.persistence;
+package talium.system.twitchCommands.persistence;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
