+++ src/main/talium/system/twitchCommands/persistence/TriggerRepo.java
@@ -1,4 +1,4 @@
-package talium.system.chatTrigger.persistence;
+package talium.system.twitchCommands.persistence;
 
 import org.springframework.data.repository.ListCrudRepository;
